refactor: mCPサーバーのコードをリファクタリング
- `src/mcp` ディレクトリを `src/mcp_servers` にリネームし、JSONPlaceholder MCPサーバーのコードを移動しました。
- `src/mcp/jsonplaceholder` ディレクトリと関連ファイル (`config.py`, `server.py`) を削除しました。
- `pyproject.toml` の依存関係とスクリプトを更新しました。
- テストコードを更新し、新しいディレクトリ構造を反映しました。
