+++ pyproject.toml
@@ -21,6 +21,7 @@ dependencies = [
     "langgraph>=0.5.0",
     "langgraph-cli[inmem]>=0.3.3",
     "langgraph-sdk>=0.1.72",
+    "mcp[cli]>=1.10.1",
     "pydantic-settings>=2.10.1",
     "pytest>=8.0.0",
     "pytest-asyncio>=0.23.0",
@@ -28,16 +29,19 @@ dependencies = [
 ]
 
 [project.scripts]
-example-agent = "src.graphs.example_agent:main"
-graph-server = "src.app:main"
-mcp-server-jsonplaceholder = "src.mcp.jsonplaceholder.server:main"
+graph-server = "app:main"
+mcp-server-jsonplaceholder = "mcp_servers.jsonplaceholder:main"
 
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src"]
+packages = [
+    "src/mcp_servers/jsonplaceholder",
+    "src/graphs",
+    "src/app"
+]
 
 [tool.pytest.ini_options]
 addopts = "-p no:warnings"
